Add files via upload
Uploading new file to use as main branch, as old file wasn't uploaded as an html file. Still in the process of getting familiar with Github, so the current process isn't the cleanest. Intention is to make this file the main branch to work from. Let's see if I can make it work.